Created currency controller
